+++ packages/ui/src/form/checkbox-control/checkbox-control.tsx
@@ -22,7 +22,7 @@ const TextContainer = ( {
 /**
  * A complete checkbox field with integrated label and description.
  *
- * Visually hidden labels also expand the checkbox pointer target.
+ * Visually hidden labels also enlarge the checkbox hit target.
  */
 export const CheckboxControl = forwardRef<
 	HTMLSpanElement,
@@ -49,7 +49,7 @@ export const CheckboxControl = forwardRef<
 						ref={ ref }
 						className={ styles.checkbox }
 						{ ...checkboxProps }
-						hasExpandedHitArea={ Boolean( hideLabelFromVision ) }
+						expandHitTarget={ Boolean( hideLabelFromVision ) }
 					/>
 				</div>
 				<TextContainer isStack={ hasSupportingText }>