+++ packages/ui/src/form/primitives/checkbox/types.ts
@@ -3,11 +3,11 @@ import type { ComponentProps } from '../../../utils/types';
 
 export type CheckboxProps = ComponentProps< typeof _Checkbox.Root > & {
 	/**
-	 * Expand the pointer target without changing the visual size.
+	 * Enlarge the hit target without changing the visual size.
 	 *
 	 * Use this when there is no visible label.
 	 *
 	 * @default false
 	 */
-	hasExpandedHitArea?: boolean;
+	expandHitTarget?: boolean;
 };